refactor: make naming for base and typescript rules and configs more consistent
