+++ src/rules/typescript-strict.js
